declare that dart2js fails some 'yield*' tests
These tests assert that 'yield*' expressions in non-generator functions
should be a compile-time error. But the spec does not have yield as a
reserved word, so a statement like 'yield* e;' can be interpreted as an
expression statement multiplying the variable 'yield' by e. This is how
dart2js and the VM interpret yield outside of generator functions.

See the discussion in issue #25495

R=sigmund@google.com

Review URL: https://codereview.chromium.org/2024203002 .