Fix Windows time zone name extraction. Update API docs
On Windows, the OS provides a full name for the time zone
rather than an abbreviation. Further the string may
contain non-ASCII characters, so a conversion is
necessary. This CL updates the API docs to match the
actual behavior, and fixes the string conversion problem.

fixes #17085

R=asiva@google.com

Review URL: https://codereview.chromium.org/2069783002 .