Do not do downward inference based on the RHS of an "as" expression.
An "as" means the user is explicitly injecting a type assertion as the
value flows out of the expression, so it's a bit fishy to push that
type down into the LHS.

Doing that can cause the "as" itself to appear unnecessary if we
calculate the type of the LHS in a context where we presume the RHS
type.

Fix #26027.

R=leafp@google.com

Review URL: https://codereview.chromium.org/2081693003 .