VM: [AOT] Make sure that we inline all resolved accessor invocations.
Previously we would just replace it with a static call and hope that inliner
handles it. However sometimes it happens too late in the pipeline and inliner
misses the opportunity.

Additionally drop some dead code from the aot_optimizer that was not doing
anything useful because FlowGraphInliner::TryInlineRecognizedMethod always
returns false in the AOT mode.

R=fschneider@google.com
BUG=

Review URL: https://codereview.chromium.org/2098643003 .