References to changed classes usually don't invalidate libraries.
When a changed class with a  delta is referenced, only references to
the changed class members invalidates resolution. Otherwise references
to the class element are still valid, because we update the element
incrementally.

OTOH, when the class is subclassed or instantiated, we might have a
different set of "unimplemented members" errors. For now we invalidate
whole libraries. In the future we will refine this to keep resolution
and invalidating only errors.

R=brianwilkerson@google.com
BUG=

Review URL: https://codereview.chromium.org/2101923002 .