Avoid to mask all the NetworkRpcException
If a malformed message is received and the test fails with a
NetworkRpcException the framework ignores it.
This is still a workaround, if the exception is no thrown from inside the test,
but from a non awaited Future, it will still pass the test.
To fix it definetly, it is required to find all the non awaited rpc requests
and await for them.

R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/2093703002 .