Parser support for trailing commas (#26647).
Adds parser support for trailing commas in parameter and argument lists.

Meta-tracking issue: https://github.com/dart-lang/sdk/issues/26644

In most cases, errors were preserved (and tested) across parsers with and without the flag set.  (The exception is a formal paramater list with an extra comma, like `(a,,)` which with trailing comma support enabled only reports a missing identifier; needless to say this is captured in a test and is up for further debate.)

Amazingly (and for better or worse), turning this option on in the parser (`parseTrailingCommas = true`) produces NO errors in the analyzer and server tests.

BUG=26647
R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org/2116853002 .