Support loading dart-ext from within a package when using a package map
Fixes https://github.com/dart-lang/sdk/issues/26879

When a dart-ext is imported from within a package: library we need to fully resolve the package: to a file: (or whatever scheme:) so that we can locate the shared object for the extension file.

This CL sends this request through the service isolate so that the package: can be resolved even when a package config is being used.

R=asiva@google.com

Review URL: https://codereview.chromium.org/2146093002 .

BUG=

patch from issue 2145983004 at patchset 1 (http://crrev.com/2145983004#ps1)