fix propagated type of lambdas in strong mode
Lambda expressions were getting propagated type inference, which conflicts with the type that strong mode infers, resulting in incorrect hints.

This hasn't shown up in a user-visible way, even though the value stored in the AST was not correct. But once implicit casts are disabled in StrongTypeSystem it started showing up as a (misleading) hint message.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org/2145273002 .