Don't schedule SOURCE_KIND computation when it is know already.
This fixes the problem when after limited invalidation we put the
changed source into the unknownSourceQueue, and schedule computation of
SOURCE_KIND, which actually never happens, because it is already valid.

The limited invalidation test will follow in a sepatate CL, because
it depends on fixing the modificationTime problem I sent for review
earlier.

R=brianwilkerson@google.com
BUG=

Review URL: https://codereview.chromium.org/2155023002 .