DartWorkManager.getLibrariesContainingPart() should ask the ResultProvider for CONTAINING_LIBRARIES.
Otherwise, because we don't actually parse units for summarized sources,
we don't build this information in DartWorkManager. But we have it
in summaries.

This fixes the problem when user opens a SDK unit and does not see any semantic highlighting, navigation, etc - because we were not able to create its fully resolved AST (which is library specific, so we need to know containing libraries). Less obvious, but similar problem is that search for references to elements defined in SDK was not working, because IDEA asks search using file and offset _in the SDK_, not at the place in workspace where it is used.

R=brianwilkerson@google.com, paulberry@google.com
BUG=

Review URL: https://codereview.chromium.org/2188523004 .