Add non-null types list to StrongModeOptions
There two tests for the list: if it's empty the behaviour should be as
before; if it's not then there should be an analyzer error. The latter
test is marked as failing, as the list is not used internally yet.

Based on https://github.com/dart-lang/dev_compiler/commit/9e124ba8d47963eb378d757ed54d257617d63bee .

BUG=
R=brianwilkerson@google.com, jmesserly@google.com

Review URL: https://codereview.chromium.org/2205743002 .