fix #26990, incorrect substitution of upper bound
We were handling the "extends" clause too early, when we did not have enough information to correctly substitute it. So we tried to substitute later on, but this breaks cases where we (correctly) inferred using the argument types.

The fix is to move the "extends" clause handling to a place where we have the inferred types for our preceding type variables available.

R=leafp@google.com

Review URL: https://codereview.chromium.org/2205993002 .