Stress test for comparing analysis performed in two AnalysisContext instances with different options.
It takes a Git repository, a folder in this repository and creates two
AnalysisContext instances - with experimental and with stable
analysis options. Then for each Git revision it applies the changes
to Dart files in the folder to both contexts, analyzes and compares
analysis cache results.

The current implementation sets "finerGrainedInvalidation" to true for the experimental context.

R=brianwilkerson@google.com
BUG=

Review URL: https://codereview.chromium.org/2206783002 .