Fix memory leaks in BoringSSL secure socket implementation
After running our tests with ASAN and inspecting the
BoringSSL code, it appears that the calls indicated
in the CL increment the reference count of the
objects passed in being added to the security context.
Therefore, to avoid a leak, we have to free() the
objects to decrement the reference count so that the
memory is freed when the context is destroyed.

R=asiva@google.com

Review URL: https://codereview.chromium.org/2206233003 .