[kernel] Ensure all allocated values come with a non-zero bitmask.
Allocation of a value and bitmask input are two different constraints
in the constraint system.  However, for all values other than 'null',
there should be exactly one bitmask input and one allocation constraint.

The bitmask input was omitted in some cases in the handling of
externals, which is fixed in this CL.

BUG=
R=kustermann@google.com

Review URL: https://chromereviews.googleplex.com/481187013 .