Fix instantiateToBounds behavior in summary resynthesizer to match that of StrongTypeSystemImpl.instantiateToBounds.
Note that StrongTypeSystemImpl.instantiateToBounds misbehaves if the
bound on one type parameter refers to a later type parameter.  I'm not
sure what the intended behavior is in this case, so I've filed
dartbug.com/27072.  Once that is fixed we can refine the behavior of
the summary resynthesizer further.

R=scheglov@google.com

Review URL: https://codereview.chromium.org/2244043002 .