Make Uri._makeQuery easier to analyze
In many programs dart2js can tell that [queryParameters] is always null.

It is too hard for dart2js to infer from the previous conditions that
the last test 'query != null' must always be true. By changing the
liveness predicate for the forEach code to be a function of one
variable, we can help dart2js understand that the forEach code is dead
when queryParameters is null.

BUG=
R=lrn@google.com

Review URL: https://codereview.chromium.org/2245323004 .