Fail linking of a node if it depends on a missing listed package.
If A depends on B and B's unlinked bundle is missing, and B is in the
.packages file, then without summaries we can resolve dependency on B.
So, we have to resolve it with summaries, but we cannot.
So, we should not link A at all.

OTOH, if B is not listed, then we cannot resolve dependency on it
even without summaries. So, we can ignore missing dependency on B.

R=brianwilkerson@google.com, paulberry@google.com
BUG=

Review URL: https://codereview.chromium.org/2257033002 .