remove duplicate checking for overrides
We were checking for override errors in two places: strong/checker.dart and error_verifier.dart, so this removes one of them.

I went for skipping the ErrorVerifier errors in strong mode. That path produces better error messages, but it's not expressed in terms of the subtype relation directly, but rather encodes information about how function subtyping works. This makes it difficult to keep in sync with our typing logic. So for now it seems pragmatic to prefer the checker implementation.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org/2246293002 .