Fix bug in locateSingleElement
This bug probably did not show up much because the conditions that
cause the bug also cause unimplemented interface method warnings.

There is a degradation in code quality in mixin method bodies which I will address later.

This bug is unrelated to the recent removal of the 'mask' parameter of locateSingleElement.

R=sigmund@google.com

Review URL: https://codereview.chromium.org/2284923002 .