Build entire program with kernel for conversion to ssa.
This CL has the kernel task of the backend building the entire program.
This causes the compiler to resolve a lot more than usual, and so we need
to support many language features in the kernel -> ssa builder to compile
all of the helper functions that are included in the output.

R=sigmund@google.com

Review URL: https://codereview.chromium.org/2338093002 .