We do and test only AST based summaries.
There are 3 tests failing.

According to the Spec 'void' and 'dynamic' cannot be prefixes.
"It is a compile-time error if a built-in identifier is used as the declared name of a prefix, class, type parameter or type alias."
So, most probably these tests should be changed to test that the prefix is null.

I will look into "fail_invalid_prefix_type_parameter" in a separate CL.

R=paulberry@google.com
BUG=

Review URL: https://codereview.chromium.org/2353993002 .