kernel->ssa: Implement for-loops and while-loops
This involved splitting out the loop handling logic from the
original SSA builder and sharing it with the kernel builder.
This was just the easiest way to do this for now, but in the future
we should write the loop handler for kernel to take into account
the simplified structure that kernel provides.

R=sra@google.com

Review URL: https://codereview.chromium.org/2360673003 .