Fix stepping over await statements
When async-stepping over an await statement that follows a previous
async-step command, ensure that the synthetic breakpoint at the
beginning of the async closure is not accidentally deleted.

Removing a test that depends on minute details of the internal
implementation of async-step commands.

Adding new test that checks for multiple consecutive async-step
commands.

Note that debugger.cc seems to contain a memory leak. Breakpoint
objects are not deleted (deallocated) when RemoveBreakpoint()
is called. Will address this in a separate CL.

BUG=#27238
R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/2366463002 .