Tweak how the pkg files timestamps are created.
The acute problem this solves is that hacking on DDC generates a large
number of files into its gen/ directory. It's enough files that once you
have done that, the SDK no longer builds successfully because it chokes
on them all.

To fix that, instead of using list_files.py, I made a little
list_dart_files.py that is explicitly for Dart files and knows to
exclude gen/ and tests. This also avoids needing the sort of hairy
regexes to match different sets of files.

As far as I can tell, this works, but GYP is pretty mysterious to me,
so let me know if I got something wrong or there is a better way to do
it.

R=whesse@google.com

Review URL: https://codereview.chromium.org/2391863002 .