Fix DBC double comparisons for NaN.
Because NaN is not handled in a separate branch (like on other architetures -
e.g. branch-on-parity-flag on ia32), and every comparison with NaN returns false,
except "!=" we can't flip the branch condition in the same way for double
compares like we do for integer compares.

We could re-order blocks to always make the false-successor the fall-through, but
for now I just disable flipping the branch condition for double compares.

BUG=#27481
R=zra@google.com

Review URL: https://codereview.chromium.org/2400103002 .