Store API signature into UnlinkedUnit.
We combine these API signatures into dependency signature.
And to add a String to MD5 we encode it into byte.
And to get a String from a FlatBuffer we decode it from bytes.
Ugh...

So, let's store signatures as List<int>.
Ideally I'd like to have an annotation like @informative, but to
specify that the value is Uint8. Maybe @int_uint8 - pretty ugly, but
that guarantees that if we get better integer types in Dart, it will not
conflict with them :-)

R=brianwilkerson@google.com, paulberry@google.com
BUG=

Review URL: https://codereview.chromium.org/2461443004 .